move param processing functions to route registry class
